requirements.txt
torchinfo
protobuf==3.20
exceptiongroup
datasets
wandb
fire
sentence-transformers
numpy
imageio